requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
future==0.18.2
idna==2.9
pycparser==2.20
pyTelegramBotAPI==3.6.7
python-telegram-bot==12.5.1
requests==2.23.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
